bids_manager/requirements.txt
flask
gunicorn
psycopg2-binary
pandas
numpy
mne>=1.6
mne-bids==0.14
edfio